+++ ntex-service/src/pipeline.rs
@@ -22,10 +22,10 @@ where
 }
 
 /// Contruct new pipeline factory with one service factory.
-pub fn pipeline_factory<T, R, F>(factory: F) -> PipelineFactory<T, R>
+pub fn pipeline_factory<T, R, C, F>(factory: F) -> PipelineFactory<T, R, C>
 where
-    T: ServiceFactory<R>,
-    F: IntoServiceFactory<T, R>,
+    T: ServiceFactory<R, C>,
+    F: IntoServiceFactory<T, R, C>,
 {
     PipelineFactory {
         factory: factory.into_factory(),
@@ -174,27 +174,22 @@ impl<T: Service<R>, R> Service<R> for Pipeline<T, R> {
 }
 
 /// Pipeline factory
-pub struct PipelineFactory<T, R> {
+pub struct PipelineFactory<T, R, C = ()> {
     factory: T,
-    _t: PhantomData<R>,
+    _t: PhantomData<(R, C)>,
 }
 
-impl<T: ServiceFactory<R>, R> PipelineFactory<T, R> {
+impl<T: ServiceFactory<R, C>, R, C> PipelineFactory<T, R, C> {
     /// Call another service after call to this one has resolved successfully.
     pub fn and_then<F, U>(
         self,
         factory: F,
-    ) -> PipelineFactory<AndThenFactory<T, U, R>, R>
+    ) -> PipelineFactory<AndThenFactory<T, U, R, C>, R, C>
     where
         Self: Sized,
-        T::Config: Clone,
-        F: IntoServiceFactory<U, T::Response>,
-        U: ServiceFactory<
-            T::Response,
-            Config = T::Config,
-            Error = T::Error,
-            InitError = T::InitError,
-        >,
+        C: Clone,
+        F: IntoServiceFactory<U, T::Response, C>,
+        U: ServiceFactory<T::Response, C, Error = T::Error, InitError = T::InitError>,
     {
         PipelineFactory {
             factory: AndThenFactory::new(self.factory, factory.into_factory()),
@@ -238,7 +233,7 @@ impl<T: ServiceFactory<R>, R> PipelineFactory<T, R> {
     /// Apply transform to current service factory.
     ///
     /// Short version of `apply(transform, pipeline_factory(...))`
-    pub fn apply<U>(self, tr: U) -> PipelineFactory<ApplyTransform<U, T, R>, R>
+    pub fn apply<U>(self, tr: U) -> PipelineFactory<ApplyTransform<U, T, R, C>, R, C>
     where
         U: Transform<T::Service>,
     {
@@ -254,14 +249,14 @@ impl<T: ServiceFactory<R>, R> PipelineFactory<T, R> {
     ///
     /// Note that this function consumes the receiving pipeline and returns a
     /// wrapped version of it.
-    pub fn then<F, U>(self, factory: F) -> PipelineFactory<ThenFactory<T, U, R>, R>
+    pub fn then<F, U>(self, factory: F) -> PipelineFactory<ThenFactory<T, U, R>, R, C>
     where
         Self: Sized,
-        T::Config: Clone,
-        F: IntoServiceFactory<U, Result<T::Response, T::Error>>,
+        C: Clone,
+        F: IntoServiceFactory<U, Result<T::Response, T::Error>, C>,
         U: ServiceFactory<
             Result<T::Response, T::Error>,
-            Config = T::Config,
+            C,
             Error = T::Error,
             InitError = T::InitError,
         >,
@@ -274,7 +269,10 @@ impl<T: ServiceFactory<R>, R> PipelineFactory<T, R> {
 
     /// Map this service's output to a different type, returning a new service
     /// of the resulting type.
-    pub fn map<F, Res>(self, f: F) -> PipelineFactory<MapServiceFactory<T, F, R, Res>, R>
+    pub fn map<F, Res>(
+        self,
+        f: F,
+    ) -> PipelineFactory<MapServiceFactory<T, F, R, Res, C>, R, C>
     where
         Self: Sized,
         F: FnMut(T::Response) -> Res + Clone,
@@ -289,7 +287,7 @@ impl<T: ServiceFactory<R>, R> PipelineFactory<T, R> {
     pub fn map_err<F, E>(
         self,
         f: F,
-    ) -> PipelineFactory<MapErrServiceFactory<T, R, F, E>, R>
+    ) -> PipelineFactory<MapErrServiceFactory<T, R, C, F, E>, R, C>
     where
         Self: Sized,
         F: Fn(T::Error) -> E + Clone,
@@ -301,7 +299,10 @@ impl<T: ServiceFactory<R>, R> PipelineFactory<T, R> {
     }
 
     /// Map this factory's init error to a different error, returning a new service.
-    pub fn map_init_err<F, E>(self, f: F) -> PipelineFactory<MapInitErr<T, R, F, E>, R>
+    pub fn map_init_err<F, E>(
+        self,
+        f: F,
+    ) -> PipelineFactory<MapInitErr<T, R, C, F, E>, R, C>
     where
         Self: Sized,
         F: Fn(T::InitError) -> E + Clone,
@@ -313,7 +314,7 @@ impl<T: ServiceFactory<R>, R> PipelineFactory<T, R> {
     }
 }
 
-impl<T, R> Clone for PipelineFactory<T, R>
+impl<T, R, C> Clone for PipelineFactory<T, R, C>
 where
     T: Clone,
 {
@@ -325,8 +326,7 @@ where
     }
 }
 
-impl<T: ServiceFactory<R>, R> ServiceFactory<R> for PipelineFactory<T, R> {
-    type Config = T::Config;
+impl<T: ServiceFactory<R, C>, R, C> ServiceFactory<R, C> for PipelineFactory<T, R, C> {
     type Response = T::Response;
     type Error = T::Error;
     type Service = T::Service;
@@ -334,7 +334,7 @@ impl<T: ServiceFactory<R>, R> ServiceFactory<R> for PipelineFactory<T, R> {
     type Future = T::Future;
 
     #[inline]
-    fn new_service(&self, cfg: T::Config) -> Self::Future {
+    fn new_service(&self, cfg: C) -> Self::Future {
         self.factory.new_service(cfg)
     }
 }