+++ ntex-service/CHANGES.md
@@ -2,7 +2,9 @@
 
 ## [0.3.0] - 2021-12-xx
 
-* Service now takes the the request type as a type parameter instead of an associated type
+* Service takes request type as a type parameter instead of an associated type
+
+* ServiceFactory takes config type as a type parameter instead of an associated type
 
 ## [0.2.1] - 2021-09-17